server/api/requirements.txt
fastapi[standard]>=0.109.0
uvicorn[standard]>=0.27.0
paho-mqtt>=2.0.0
influxdb-client>=1.40.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
redis>=5.0.0
python-dotenv>=1.0.0
